+++ apps/frontend/src/_components/result/ButtonGroupSection.module.css
