+++ apps/frontend/src/_components/result/ProgressBar.tsx
