+++ apps/frontend/src/_components/result/ScoreSection.module.css
