+++ apps/frontend/src/_components/result/ScoreSection.tsx
