+++ apps/frontend/src/_components/result/TotalEvaluationSection.module.css
