+++ apps/frontend/src/_components/result/TotalEvaluationSection.tsx
